Quick change to compress and preprocess into double to avoid the cast. Hail mary and its should only be taken as a quick experiment,

+++ src/compression.cpp
@@ -27,7 +27,7 @@ unsigned long maxCompressedDataSize(const unsigned int numFloats)
     return maxOutputSize;
 }
 
-unsigned long compressData(const VectorXf &snpData, unsigned char *outputBuffer, unsigned long outputSize)
+unsigned long compressData(const VectorXd &snpData, unsigned char *outputBuffer, unsigned long outputSize)
 {
     // Initialise zlib
     z_stream strm;
@@ -40,9 +40,9 @@ unsigned long compressData(const VectorXf &snpData, unsigned char *outputBuffer,
         return 0;
 
     // Compress the data
-    const unsigned int inputSize = static_cast<unsigned int>(snpData.size()) * sizeof(float);
+    const unsigned int inputSize = static_cast<unsigned int>(snpData.size()) * sizeof(double);
     strm.avail_in = inputSize;
-    strm.next_in = reinterpret_cast<unsigned char *>(const_cast<float*>(&snpData[0]));
+    strm.next_in = reinterpret_cast<unsigned char *>(const_cast<double*>(&snpData[0]));
     strm.avail_out = static_cast<unsigned int>(outputSize);
     strm.next_out = outputBuffer;
